refactor: extract asset impl
